fix: リハーサル当日なのに実装ミスらないで？？？？

+++ main.py
@@ -31,22 +31,22 @@ def get(playlist_id, key, directory):
     api = yt_api.YT_API(key)
     pitems = api.playlist_items(playlist_id)["items"]
     ids = list(map(lambda i: i["contentDetails"]["videoId"], pitems))
-
     ext = "mkv"
     with open(f"{directory}/list.m3u8", "a", encoding="UTF-8") as f:
         for i in split_by_fifty(ids):
             v = api.videos(",".join(i))["items"]
             for j in v:
-                url = f"https://youtu.be/{j['id']}"
-                filename = f"{j['snippet']['title']} - {j['snippet']['channelTitle']}"
-                subprocess.call(["yt-dlp", url, "--merge-output-format", ext, "-o", f'{directory}/{filename}.%(ext)s'])
-                if(os.path.exists(f"{directory}/{filename}.mp4")):
-                    subprocess.call(["ffmpeg", "-i", f"{directory}/{filename}.mp4", "-c:v", "copy", "-c:a", "copy", f"{directory}/{filename}.mkv"])
-                    os.remove(f"{directory}/{filename}.mp4")
-                elif(os.path.exists(f"{directory}/{filename}.webm")):
-                    subprocess.call(["ffmpeg", "-i", f"{directory}/{filename}.webm", "-c:v", "copy", "-c:a", "copy", f"{directory}/{filename}.mkv"])
-                    os.remove(f"{directory}/{filename}.webm")
-                f.write(f"{filename}.{ext}\n")
+                if("error" in v):
+                    url = f"https://youtu.be/{j['id']}"
+                    filename = f"{j['snippet']['title']} - {j['snippet']['channelTitle']}"
+                    subprocess.call(["yt-dlp", url, "--merge-output-format", ext, "-o", f'{directory}/{filename}.%(ext)s'])
+                    if(os.path.exists(f"{directory}/{filename}.mp4")):
+                        subprocess.call(["ffmpeg", "-i", f"{directory}/{filename}.mp4", "-c:v", "copy", "-c:a", "copy", f"{directory}/{filename}.mkv"])
+                        os.remove(f"{directory}/{filename}.mp4")
+                    elif(os.path.exists(f"{directory}/{filename}.webm")):
+                        subprocess.call(["ffmpeg", "-i", f"{directory}/{filename}.webm", "-c:v", "copy", "-c:a", "copy", f"{directory}/{filename}.mkv"])
+                        os.remove(f"{directory}/{filename}.webm")
+                    f.write(f"{filename}.{ext}\n")
 
 def main():
     if(len(sys.argv) == 4):
